refactor: move files about seat selection to reservation package
