+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.SeatSelectionActivity"
+            android:name=".view.reservation.SeatSelectionActivity"
             android:exported="false" />
         <activity
             android:name=".view.reservation.ReservationActivity"
