+++ app/src/main/java/woowacourse/movie/domain/reservation/Column.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.reservation
 
 @JvmInline
 value class Column(
