+++ app/src/main/java/woowacourse/movie/domain/reservation/Row.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.reservation
 
 @JvmInline
 value class Row(
