+++ app/src/main/java/woowacourse/movie/domain/reservation/Seat.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.reservation
 
 import java.io.Serializable
