+++ app/src/main/res/layout/activity_seat_selection.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SeatSelectionActivity">
+    tools:context=".view.reservation.SeatSelectionActivity">
 
     <ScrollView
         android:layout_width="match_parent"
